+++ src/crosssections.cc
@@ -2597,7 +2597,7 @@ CollisionBranchList CrossSections::string_excitation(
           finder_parameters.hard_string_transition_end_energy;
 
       const double weight_hard =
-          smooth_transition_weight(hard_transition_start, hard_transition_end);
+          interpolation_at_sqrts(hard_transition_start, hard_transition_end);
 
       nondiffractive_hard = nondiffractive_all * weight_hard;
       nondiffractive_soft = nondiffractive_all - nondiffractive_hard;
@@ -2676,7 +2676,7 @@ double CrossSections::high_energy(
        * that defined in string_probability(). */
       auto [region_lower, region_upper] =
           finder_parameters.transition_high_energy.sqrts_range_NN;
-      double prob_high = probability_transit_high(region_lower, region_upper);
+      double prob_high = interpolation_at_sqrts(region_lower, region_upper);
       xs = xs_l * (1. - prob_high) + xs_h * prob_high;
     }
   }
@@ -3119,13 +3119,13 @@ double CrossSections::string_probability(
       return 0.;
     } else {
       // Rescale transition region to [-1, 1]
-      return probability_transit_high(region_lower, region_upper);
+      return interpolation_at_sqrts(region_lower, region_upper);
     }
   }
 }
 
-double CrossSections::smooth_transition_weight(double region_lower,
-                                               double region_upper) const {
+double CrossSections::interpolation_at_sqrts(double region_lower,
+                                             double region_upper) const {
   if (sqrt_s_ < region_lower)
     return 0.0;
   if (sqrt_s_ > region_upper)
@@ -3138,9 +3138,5 @@ double CrossSections::smooth_transition_weight(double region_lower,
   assert(prob >= 0. && prob <= 1.);
 
   return prob;
-}
-double CrossSections::probability_transit_high(
-    const double region_lower, const double region_upper) const {
-  return smooth_transition_weight(region_lower, region_upper);
 }
 }  // namespace smash